ui: Hoist SplitButton out of git_ui into ui
