+++ crates/ui/src/components/button.rs
@@ -2,9 +2,11 @@ mod button;
 mod button_icon;
 mod button_like;
 mod icon_button;
+mod split_button;
 mod toggle_button;
 
 pub use button::*;
 pub use button_like::*;
 pub use icon_button::*;
+pub use split_button::*;
 pub use toggle_button::*;
